FIX: Remove warning.

+++ Packages/com.unity.inputsystem/InputSystem/Events/InputEvent.cs
@@ -76,7 +76,7 @@ namespace UnityEngine.Experimental.Input.LowLevel
         public int eventId
         {
             get => (int)(m_Event.eventId & kIdMask);
-            set => m_Event.eventId = (int)(value | (m_Event.eventId & ~kIdMask));
+            set => m_Event.eventId = (int)(value | (int)(m_Event.eventId & ~kIdMask));
         }
 
         /// <summary>
